Task2

+++ Tyuiu.GalchinskayaVV.Sprint1.Task2.V21.Lib/DataService.cs
@@ -1,12 +1,13 @@
 ﻿using tyuiu.cources.programming.interfaces.Sprint1;
+using System;
 
 namespace Tyuiu.GalchinskayaVV.Sprint1.Task2.V21.Lib
 {
     public class DataService : ISprint1Task2V21
     {
-        public double Calculate(double x, double y)
+        public int Calculate(double x, double y)
         {
-            return x * y;
+            return (int)(x * y);
         }
 
         public int CalculateRectangleSquare(int value, int valueTwo)
